+++ flask_via/examples/small/foo/routes.py
@@ -8,8 +8,15 @@ A small ``Flask-Via`` example Flask application.
 """
 
 from flask_via.examples.small.foo import views
-from flask.ext.via.routers import default
+from flask.ext.via.routers import default, Include
 
 routes = [
     default.Basic('/foo', views.foo),
+    Include('flask_via.examples.small.foo.routes',
+            routes_name='urls',
+            url_prefix='/bar')
+]
+
+urls = [
+    default.Basic('/bar', views.foo),
 ]