+++ flask_via/examples/small/routes.py
@@ -15,5 +15,5 @@ routes = [
     default.Basic('/about', views.about),
     default.Basic('/contact', views.contact),
     # Include other routes from other modules
-    Include('flask_via.examples.small.foo.routes')
+    Include('flask_via.examples.small.foo.routes', url_prefix='/foo')
 ]